Structure Refactor
net.hydrius.join -> net.hydrius.hydriusjoin

This is done for consistency.
