+++ src/main/java/net/hydrius/hydriusjoin/util/group/ItemGroup.java
@@ -1,9 +1,9 @@
-package net.hydrius.join.util.group;
+package net.hydrius.hydriusjoin.util.group;
 
 import dev.flrp.espresso.util.EnchantUtils;
 import dev.flrp.espresso.util.StringUtils;
-import net.hydrius.join.util.action.Action;
-import net.hydrius.join.util.action.ActionUtils;
+import net.hydrius.hydriusjoin.util.action.Action;
+import net.hydrius.hydriusjoin.util.action.ActionUtils;
 import org.bukkit.Material;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.inventory.ItemStack;
